feat: add support for new POST zones/:zone/verify endpoint
Dennis now has support for verifying a Zone using a TXT record.
This in addition to the existing nameserver based verification.

The new POST zones/:zone/verify endpoint will check both verification methods.

This PR adds a new `zone#verify` method to hit this endpoint.

It also exposes two new Zone attributes which are returned in the API response:
- `txt_record_verification_token`
- `verified_at`

In addition this PR bumps Rubocop to the latest version and adds binstubs to make it easier to run the linting and specs via:

- bin/rubocop
- bin/rspec spec